refactor(type): split signal feedback engage types
