+++ Source/Portfolio/Component/CCombatSignalSourceComponent.h
@@ -3,7 +3,7 @@
 #include "CoreMinimal.h"
 #include "Components/ActorComponent.h"
 #include "Type/CCharacterComponentReferenceTypes.h"
-#include "Type/CCombatSignalStructure.h"
+#include "Type/CCombatSignalTypes.h"
 #include "Type/CWeaponStructure.h"
 #include "CCombatSignalSourceComponent.generated.h"
