+++ Source/Portfolio/Component/CCombatSignalTargetComponent.h
@@ -3,7 +3,7 @@
 #include "CoreMinimal.h"
 #include "Components/ActorComponent.h"
 #include "Type/CCharacterComponentReferenceTypes.h"
-#include "Type/CCombatSignalStructure.h"
+#include "Type/CCombatSignalTypes.h"
 #include "Type/CWeaponStructure.h"
 #include "CCombatSignalTargetComponent.generated.h"
