+++ Source/Portfolio/Core/Debug/FCombatEngageDebug.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "Type/CWorldSubsystemStructure.h"
+#include "Type/CEngageAssignmentTypes.h"
 
 class AActor;
 class ACAIController;
