+++ Source/Portfolio/System/Combat/CWorldSubsystem_CombatEngage.cpp
@@ -10,7 +10,7 @@
 #include "Core/Profiling/CCombatCollisionProfilingCounters.h"
 #include "Core/Profiling/CCombatFeedbackProfiling.h"
 
-#include "Type/CWorldSubsystemStructure.h"
+#include "Type/CEngageAssignmentTypes.h"
 
 namespace
 {
