+++ Source/Portfolio/System/Combat/CWorldSubsystem_CombatEngage.h
@@ -2,7 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "Subsystems/WorldSubsystem.h"
-#include "Type/CWorldSubsystemStructure.h"
+#include "Type/CEngageAssignmentTypes.h"
 #include "CWorldSubsystem_CombatEngage.generated.h"
 
 UCLASS()
