--- Source/Portfolio/Type/CCombatSignalStructure.cpp
@@ -1,1 +1,0 @@
-#include "Type/CCombatSignalStructure.h"
