--- Source/Portfolio/Type/CWorldSubsystemStructure.cpp
@@ -1,1 +1,0 @@
-#include "Type/CWorldSubsystemStructure.h"
